Separate cli command construction into new file
